clean, mypy, flake8

+++ src/farmer.py
@@ -274,11 +274,11 @@ class Farmer:
             # TODO: ask the pool for this information
 
             pool_sks: List[PrivateKey] = [
-                PrivateKey.from_bytes(bytes.fromhex(ce))
+                PrivateKey.from_bytes(bytes.fromhex(ce)) # type: ignore # noqa
                 for ce in self.key_config["pool_sks"]
             ]
 
-            coinbase_reward = uint64(calculate_block_reward(proof_of_space_finalized.height) / 8) * 7
+            coinbase_reward = uint64(int((calculate_block_reward(proof_of_space_finalized.height) / 8) * 7))
             coinbase_coin, coinbase_signature = create_coinbase_coin_and_signature(
                 proof_of_space_finalized.height + 1,
                 bytes.fromhex(self.key_config["pool_target"]),
